CommonLawCratsBackend/requirements.txt
django==4.2.5
requests
djangorestframework==3.14.0
streamlit==1.24.1
PyMuPDF==1.24.10
channels==4.0.0
faiss-cpu==1.7.4
openai
pymysql==1.1.0
tiktoken==0.4.0
mysqlclient==2.2.0
requests==2.31.0
scikit-learn==1.3.0
pandas==2.0.3
#dont need below three package -> see AllLegalMLTools.views.py and AllLegalMLTools.case_summariser.py
# langchain_community==0.0.1
# langchain_openai==0.1.8
# langchain-core==0.0.14
transformers==4.31.0
python-dotenv==1.0.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.2.2
django-allauth==0.57.0
cryptography==41.0.3
django-rest-framework-social-oauth2==1.2.0
gunicorn==21.2.0
django-environ==0.9.0
